Added codes for flushing loaded handles before the resource manager is started.
tboot 1.8.3(TPM 2.0) create an object handle during boot process.
However, the resource manager of TSS seems to presume this handle as
abnormal object and returns an error code to client.

To solve this problems, the resource manager is changed to flush
all loaded handles before the resource manager is started.

This commit solved issue #83