build: Move all distributed headers into a common directory.
The end goal of this work is to make the include directives in our
source code consistent with the layout that consuming applications will
have. This isn't so important for the library code, but for the test /
client / common code we've built up it is. This is the first step in a
larger effort to cleanup the compiler search paths we've had to set in
the build and to ease some of the cleanup that needs to be done over in
the tpm2.0-tools repo.

This is part of issue #93.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>