tcti_socket: Remove simulator flag.
The resourcemgr is a better place to filter PlatformCommands. The TCTI
shouldn't care whether or not the resourcemgr is talking to the
simulator or not. We effectively move the check up a level in the call
scope to the resource manager. The check now happens before the resource
manager calls PlatformCommand.

Consuming applications will no longer be required to define this
variable. The socket TCTI will always send platform commands when the
command socket is connected.

This resolves #131

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>