debug: Stop using global printfFunction pointer in debug functions.
We also remove the extern declaration of the same function pointer in the
debug header since it's not defined in the debug implementation. This
requires that we move the declaration to the socket TCTI header. This is
a bit ugly as it exposes this global data to clients but at least it's
not misleading. The enumeration we use as the second parameter to this
function must be moved to the socket TCTI header as well.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>