debug: Remove global variable rmDebugPrefix.
This global variable seemed to act like an escape hatch overriding the
flag passed to the debug function. As someone coming along and trying to
pick up this code I found it confusing when debug function refused to
print a prefix even though I had asked it to. Not hard to figure out but
global flags like this have the potential to cause unexpected behavior.
IMHO it's better to simplify the code and remove this.

To compensate for the loss of the rmDebugPrefix flag in the
DebugPrintBuffer function I've added another parameter for callers to pass
a printf_type flag to obtain the same behavior as when the rmDebugPrefix
flag was set.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>