test: Add test cases to test TCTI logging mechanism in device TCTI.
I had to break my own goal of keeping the cmocka specific code in the
test driver and out of the test functions. Their assert_* functions are
super useful through and much better than a pile of if / else tests on
return values.
I'm also not 100% convinced I like this structure for the test case
layout but it's not too late to change this up if we can come up with a
better approach.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>