Remove debug data from TCTI function signature.
Both the device and socket TCTI init functions take as a final parameter
a string. AFAIK these strings are only used in a debug print function as
a way to track which TCTI is being initialized and when. This is useful
to client code but not something that should be reflected in the API
that we expose.

This patch removes this parameter and the corresponding debug printf
call from both TCTI init functions. An identical debug printf is added
to each client (resourcemgr, tpmclient, tpmtest) to keep the output
consistent. Unit tests are updated as well.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>