sockets: Remove global data from InitSockets.
This function is used by both the socket TCTI and the resource manager
to initialize socket stuff. The use of the printfFunction pointer here
creates some tight coupling between resource manager and the tcti_socket
code. To keep the debug output from this function unchanged we add two
new parameters, the first is a function pointer to a debug function and
the second is a pointer to some user data. This is the same function
signature as is used by the TCTI for the user debug callback function.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>