tss2_tcti.h: Add macros to simplify access to data in the common TCTI structure.
Accessing data in the common TCTI structure requires some acrobatics.
This takes the shape of a cast, some dereferencing and a pile of
parentheses. This makes simple things like a function invocation very
difficult to read. What's not as obvious is the safety of such
operations because it's easy to forget to check for error conditions.

The spec offers an example macro for invoking functions from the common
TCTI structure. This code uses that macro, adding support for the other
functions.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>