getcommands: Remove use of malloc function pointer.
This function is only called from the resourcemgr so we really only ever
need the RM specific function pointer. Further, we have no use case to
provide a different allocation function in the RM so this only
complicates the build and makes the code more difficult to read.

Likely this was inteded to be a "future proofing" mechanism in the event
that the RM would be built on systems with different allocation
mechanisms?

Signed-off-by: Philip Tricca <flihp@twobit.us>