Tss2_Sys_Initialize: Factor the initialization of data buffer pointers into a separate function (InitSysContextPtrs).
These pointers are only initialized once so it made a certain amount of
sense to just do this in the main context initialization function. But
when we start integrating unit tests for the marshalling functions it's
nice to have a function to call that short cuts all of the consistency
checks in Tss2_Sys_Initialize. With this simple function we can
initialize all of the pointers to the data regions where data is
marshalled and then just call the relevant marshalling functions without
having to do a pile of setup.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>