Add example systemd service file for resourcemgr and udev rules for tpm device node.
These work as-is on Debian provided there's a system user named 'tss'
already and the PREFIX is '/usr/local'. Integrating these files into
the build shouldn't be too hard but the only systemd module for autoconf
that I can find is out of tree and not in the archive either.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>