tcti_socket: Remove obsolete comments.
Uncommending a #define as part of our debug instructions is wrong. The
right way to do this is to use the debug_config.site file with the
CONFIG_SITE environment variable @ ./configure time.

Also the bit about how the socket TCTI is built differently for the RM
and library use is no longer accurate.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>